Structure model.py to be an encapsulating class.
- Moved all data loading/pre-processing to driver.py script
- Updated README.md to have correct usage instructions
- Added some comments
